split up snapshot logic into submodules
