+++ packages/core/lib/v3/types/private/index.ts
@@ -4,3 +4,4 @@ export * from "./internal";
 export * from "./evaluator";
 export * from "./cache";
 export * from "./agent";
+export * from "./snapshot";
