+++ packages/core/lib/v3/understudy/a11y/snapshot/index.ts
@@ -1,0 +1,4 @@
+export { captureHybridSnapshot } from "./capture";
+export { computeActiveElementXpath } from "./activeElement";
+export { diffCombinedTrees } from "./treeFormatUtils";
+export { resolveXpathForLocation } from "./coordinateResolver";
